pyproject.toml
[project]
name = "narrate"
version = "0.1.0"
description = "A narration tool"
authors = [
    {name = "Mihai Cosma", email = "mcosma@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "phonemizer",
    "torch",
    "transformers",
    "scipy",
    "munch",
    "python-dotenv",
    "requests",
    "openai",  # For Recraft API compatibility
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["narrate*"]
namespaces = false

[tool.setuptools.package-data]
"*" = ["*.txt", "*.json"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.pylint.format]
max-line-length = "200"

[tool.ruff]
# Default is: pycodestyle (E) and Pyflakes (F)
# We add flake8-builtins (A), pydocstyle (D), isort (I), pep8-naming (N), and pylint (PL).
# We remove pycodestyle (E) since it throws erroneous line too long errors.
# We remove Pyflakes (F) since it complains about `import *` which we need.
select = ["A", "D", "I", "N", "PL"]

# We ignore the following rules:
# A002: Argument `X` is shadowing a Python builtin
# D100: missing docstring in public module
# D101: missing docstring in public class
# D102: missing docstring in public method
# D103: missing docstring in public function
# D105: missing docstring in magic method
# D107: Missing docstring in `__init__`
# D203: 1 blank line required before class docstring (incompatible with D211: no blank lines before class docstring)
# D213: multi-line-summary-second-line (incompatible with D212: multi-line summary should start at the first line)
# D413: Missing blank line after last section
# D415: First line should end with a period, question mark, or exclamation point (in period-only D400)
# D416: section-name-ends-in-colon (numpy style guide doesn't use colons after sections, i.e. Parameters)
# PLR0912: Too many branches (X > 12)
# PLR0913: Too many arguments to function call (X > 5)
# PLR0915: Too many statements (X > 50)
# PLR2004: Magic value used in comparison
ignore = ["A002", "D100", "D101", "D102", "D103", "D105", "D107", "D203", "D213", "D413", "D415", "D416", "PLR0912", "PLR0913", "PLR0915", "PLR2004"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "D", "I", "N", "PL"]
unfixable = []

line-length = 200

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Tested with Python 3.13
target-version = "py313"
